+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 1.1.3 (2015-03-13)
 
 * Added ``fluent_utils.django_compat.get_app_label()``
 * Added ``fluent_utils.dry.models.get_db_table()``
+* Added ``fluent_utils.dry.fields.HideChoicesCharField``
 
 
 Version 1.1.2 (2015-01-08)